+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [4.7.2] - 2016-27-02
+- Added option to toggle JBLM marker on/off.
+
+## [4.7.1] - 2016-28-02
+### Fixed
+- timestamp issue
+
+## [4.7.0] - 2016-21-02
 ### Added
 - Refresh button to pass report page.
 - Tip View for traffic map overflow menu.